// display-manager/src/app/common/models/general-codes-model.ts
export enum EventStatus {
    Suspended = "SUSPENDED",
    Active = "ACTIVE"
}

export enum MarketStatus {
    Suspended = "SUSPENDED",
    Active = "ACTIVE"
}

export enum SelectionStatus {
    Suspended = "SUSPENDED",
    Active = "ACTIVE"
}

export enum DisplayStatus {
    NotDisplayed = "NOTDISPLAYED",
    Displayed = "DISPLAYED",
}

export enum PriceType {
    startPrice = "SP",
    nonRunner = "N/R",
    evs = "EVS"
}

export enum SelectionSuspended {
    selectionAndPrice = ""
}

export enum EachWayTerm{
    WinOnly = 'Win Only',
    Places =' places ',
    Odds = ' odds',
    Ratio = '/',
}

export enum HorseRacing {
    Reserve = "(RES)",
    NonRunner = "NON-RUNNER",
    SpotLight = "spotlight"
}